feat: cashout work
